+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 - Breaking change to the `Documentation` constructor argument (object):
   - property `hasSummaryFromDescription` (boolean) replaced with `summarizer` (function);
   - If used with `false` value, replace it with `summarizer: ({ summary, trim }) => trim(summary)` for same behavior;
-- Featuring `summarizer` option to customize the summary of the Endpoint in the genedated Documentation:
+- Featuring `summarizer` option to customize the summary of the Endpoint in the generated Documentation:
   - The function receives `summary`, `description` and the default `trim()` function as arguments;
   - The default summarizer uses `description` as a fallback for missing `summary`;
   - The `trim()` function accepts a string and the limit (default: 50, best practice) that you can now customize;