+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Version 27
 
+### v27.2.2
+
+- `@express-zod-api/zod-plugin` version bumped to `^4.1.0`.
+
 ### v27.2.1
 
 - Improved handling of native Express middleware return values: